TryHackMe path and module descriptions added
Added descriptions for the paths and modules of TryHackMe courses for certificates of completion. Pre Security path is complete, and the certificate was added to the repo in a previous commit. The Jr Penetration Tester path is currently underway (24% complete as of the time of this commit...) and the certificate will be added to the repo once completed. - AF